+++ projects/swimlane/ngx-ui/src/lib/components/plus-menu/plus-menu-position.enum.ts
@@ -1,0 +1,5 @@
+export enum PlusMenuPosition {
+  Right = 'right',
+  Bottom = 'bottom',
+  Top = 'top'
+}