+++ projects/swimlane/ngx-ui/src/lib/services/injection/injection.service.ts
@@ -129,7 +129,7 @@ export class InjectionService {
   }
 
   /**
-   * Appends a component to a adjacent location
+   * Appends a component to an adjacent location
    *
    * @param componentClass
    * @param [options={}]