+++ projects/swimlane/ngx-ui/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 # HEAD (unreleased)
 
+- Feature(Stepper): Add `[large]` input to increase the default icon size.
+- Feature(Stepper): Add `[trackBar]` input to make track bar optional.
+- Feature(Stepper): Add `[progress]` input to show a progress indicator on the active step.
+- Feature(Stepper): Add `[icon]` input to show an icon instead of the step number.
+- Feature(Stepper): Add `.complete()` method to complete every step (including the last one).
+
 ## 32.0.0 (2020-11-24)
 
 - Enhancement: ngx card component