+++ projects/swimlane/ngx-ui/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 # HEAD (unreleased)
 
+- Feature: add `ngx-plus-menu`
 - Enhancement: Add `required` and `requiredIndicator` inputs to `DateTimeComponent`.
 - Breaking: An empty value will no longer cause the `DateTimeComponent` to become invalid, unless it is explicitly marked as `required`.