+++ projects/swimlane/ngx-ui/CHANGELOG.md
@@ -5,6 +5,12 @@
 - Feature: add `ngx-plus-menu`
 - Enhancement: Add `required` and `requiredIndicator` inputs to `DateTimeComponent`.
 - Breaking: An empty value will no longer cause the `DateTimeComponent` to become invalid, unless it is explicitly marked as `required`.
+- Feature(Stepper): Add `[large]` input to increase the default icon size.
+- Feature(Stepper): Add `[trackBar]` input to make track bar optional.
+- Feature(Stepper): Add `[progress]` input to show a progress indicator on the active step.
+- Feature(Stepper): Add `[removeHighlight]` input to remove highlight color for the completed steps.
+- Feature(Stepper): Add `[icon]` input to show an icon instead of the step number.
+- Feature(Stepper): Add `.complete()` method to complete every step (including the last one).
 
 ## 32.0.0 (2020-11-24)